add id autoincremental

+++ src/main/java/com/backend/spring/website/api/repository/WebSiteRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 
 import com.backend.spring.website.api.model.WebSite;
 
-public interface WebSiteRepository extends MongoRepository<WebSite, Long>{
+public interface WebSiteRepository extends MongoRepository<WebSite, Integer>{
 
 }
